requirements.txt
streamlit
openai
newspaper3k
lxml_html_clean